backend: replace MYSQL with SQLite
SQLite is an embedded databese that does not require a separate
server.
Using SQLite in place of a dedicated database server reduces the
complexity of the backend stack which in turn makes it easier for
developers to work with the system. This better reflects the current
state of the system, where no enterprise server environment is
necessary.